Refactor master slave integrate test cases 13th.
