+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/expect/select/SelectEqualsWithSingleMasterTable_0.xml
@@ -1,3 +1,3 @@
 <dataset>
-    <t_order order_id="1000" user_id="10" status="init_master" />
+    <t_order order_id="1000" user_id="10" status="init" />
 </dataset>
