+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/init/master_0.xml
@@ -1,37 +1,37 @@
 <dataset>
-    <t_order_0 order_id="1000" user_id="10" status="init_master" />
-    <t_order_1 order_id="1001" user_id="10" status="init_master" />
-    <t_order_2 order_id="1002" user_id="10" status="init_master" />
-    <t_order_3 order_id="1003" user_id="10" status="init_master" />
-    <t_order_4 order_id="1004" user_id="10" status="init_master" />
-    <t_order_5 order_id="1005" user_id="10" status="init_master" />
-    <t_order_6 order_id="1006" user_id="10" status="init_master" />
-    <t_order_7 order_id="1007" user_id="10" status="init_master" />
-    <t_order_8 order_id="1008" user_id="10" status="init_master" />
-    <t_order_9 order_id="1009" user_id="10" status="init_master" />
+    <t_order_0 order_id="1000" user_id="10" status="init" />
+    <t_order_1 order_id="1001" user_id="10" status="init" />
+    <t_order_2 order_id="1002" user_id="10" status="init" />
+    <t_order_3 order_id="1003" user_id="10" status="init" />
+    <t_order_4 order_id="1004" user_id="10" status="init" />
+    <t_order_5 order_id="1005" user_id="10" status="init" />
+    <t_order_6 order_id="1006" user_id="10" status="init" />
+    <t_order_7 order_id="1007" user_id="10" status="init" />
+    <t_order_8 order_id="1008" user_id="10" status="init" />
+    <t_order_9 order_id="1009" user_id="10" status="init" />
     
-    <t_order_item_0 item_id="100000" order_id="1000" user_id="10" status="init_master" />
-    <t_order_item_0 item_id="100001" order_id="1000" user_id="10" status="init_master" />
-    <t_order_item_1 item_id="100100" order_id="1001" user_id="10" status="init_master" />
-    <t_order_item_1 item_id="100101" order_id="1001" user_id="10" status="init_master" />
-    <t_order_item_2 item_id="100200" order_id="1002" user_id="10" status="init_master" />
-    <t_order_item_2 item_id="100201" order_id="1002" user_id="10" status="init_master" />
-    <t_order_item_3 item_id="100300" order_id="1003" user_id="10" status="init_master" />
-    <t_order_item_3 item_id="100301" order_id="1003" user_id="10" status="init_master" />
-    <t_order_item_4 item_id="100400" order_id="1004" user_id="10" status="init_master" />
-    <t_order_item_4 item_id="100401" order_id="1004" user_id="10" status="init_master" />
-    <t_order_item_5 item_id="100500" order_id="1005" user_id="10" status="init_master" />
-    <t_order_item_5 item_id="100501" order_id="1005" user_id="10" status="init_master" />
-    <t_order_item_6 item_id="100600" order_id="1006" user_id="10" status="init_master" />
-    <t_order_item_6 item_id="100601" order_id="1006" user_id="10" status="init_master" />
-    <t_order_item_7 item_id="100700" order_id="1007" user_id="10" status="init_master" />
-    <t_order_item_7 item_id="100701" order_id="1007" user_id="10" status="init_master" />
-    <t_order_item_8 item_id="100800" order_id="1008" user_id="10" status="init_master" />
-    <t_order_item_8 item_id="100801" order_id="1008" user_id="10" status="init_master" />
-    <t_order_item_9 item_id="100900" order_id="1009" user_id="10" status="init_master" />
-    <t_order_item_9 item_id="100901" order_id="1009" user_id="10" status="init_master" />
+    <t_order_item_0 item_id="100000" order_id="1000" user_id="10" status="init" />
+    <t_order_item_0 item_id="100001" order_id="1000" user_id="10" status="init" />
+    <t_order_item_1 item_id="100100" order_id="1001" user_id="10" status="init" />
+    <t_order_item_1 item_id="100101" order_id="1001" user_id="10" status="init" />
+    <t_order_item_2 item_id="100200" order_id="1002" user_id="10" status="init" />
+    <t_order_item_2 item_id="100201" order_id="1002" user_id="10" status="init" />
+    <t_order_item_3 item_id="100300" order_id="1003" user_id="10" status="init" />
+    <t_order_item_3 item_id="100301" order_id="1003" user_id="10" status="init" />
+    <t_order_item_4 item_id="100400" order_id="1004" user_id="10" status="init" />
+    <t_order_item_4 item_id="100401" order_id="1004" user_id="10" status="init" />
+    <t_order_item_5 item_id="100500" order_id="1005" user_id="10" status="init" />
+    <t_order_item_5 item_id="100501" order_id="1005" user_id="10" status="init" />
+    <t_order_item_6 item_id="100600" order_id="1006" user_id="10" status="init" />
+    <t_order_item_6 item_id="100601" order_id="1006" user_id="10" status="init" />
+    <t_order_item_7 item_id="100700" order_id="1007" user_id="10" status="init" />
+    <t_order_item_7 item_id="100701" order_id="1007" user_id="10" status="init" />
+    <t_order_item_8 item_id="100800" order_id="1008" user_id="10" status="init" />
+    <t_order_item_8 item_id="100801" order_id="1008" user_id="10" status="init" />
+    <t_order_item_9 item_id="100900" order_id="1009" user_id="10" status="init" />
+    <t_order_item_9 item_id="100901" order_id="1009" user_id="10" status="init" />
     
-    <t_config id="1" status="init_master" />
+    <t_config id="1" status="init" />
     
     <t_global id="1" status="global_master" />
 </dataset>
