+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/init/master_1.xml
@@ -1,35 +1,35 @@
 <dataset>
-    <t_order_0 order_id="1100" user_id="11" status="init_master" />
-    <t_order_1 order_id="1101" user_id="11" status="init_master" />
-    <t_order_2 order_id="1102" user_id="11" status="init_master" />
-    <t_order_3 order_id="1103" user_id="11" status="init_master" />
-    <t_order_4 order_id="1104" user_id="11" status="init_master" />
-    <t_order_5 order_id="1105" user_id="11" status="init_master" />
-    <t_order_6 order_id="1106" user_id="11" status="init_master" />
-    <t_order_7 order_id="1107" user_id="11" status="init_master" />
-    <t_order_8 order_id="1108" user_id="11" status="init_master" />
-    <t_order_9 order_id="1109" user_id="11" status="init_master" />
+    <t_order_0 order_id="1100" user_id="11" status="init" />
+    <t_order_1 order_id="1101" user_id="11" status="init" />
+    <t_order_2 order_id="1102" user_id="11" status="init" />
+    <t_order_3 order_id="1103" user_id="11" status="init" />
+    <t_order_4 order_id="1104" user_id="11" status="init" />
+    <t_order_5 order_id="1105" user_id="11" status="init" />
+    <t_order_6 order_id="1106" user_id="11" status="init" />
+    <t_order_7 order_id="1107" user_id="11" status="init" />
+    <t_order_8 order_id="1108" user_id="11" status="init" />
+    <t_order_9 order_id="1109" user_id="11" status="init" />
     
-    <t_order_item_0 item_id="110000" order_id="1100" user_id="11" status="init_master" />
-    <t_order_item_0 item_id="110001" order_id="1100" user_id="11" status="init_master" />
-    <t_order_item_1 item_id="110100" order_id="1101" user_id="11" status="init_master" />
-    <t_order_item_1 item_id="110101" order_id="1101" user_id="11" status="init_master" />
-    <t_order_item_2 item_id="110200" order_id="1102" user_id="11" status="init_master" />
-    <t_order_item_2 item_id="110201" order_id="1102" user_id="11" status="init_master" />
-    <t_order_item_3 item_id="110300" order_id="1103" user_id="11" status="init_master" />
-    <t_order_item_3 item_id="110301" order_id="1103" user_id="11" status="init_master" />
-    <t_order_item_4 item_id="110400" order_id="1104" user_id="11" status="init_master" />
-    <t_order_item_4 item_id="110401" order_id="1104" user_id="11" status="init_master" />
-    <t_order_item_5 item_id="110500" order_id="1105" user_id="11" status="init_master" />
-    <t_order_item_5 item_id="110501" order_id="1105" user_id="11" status="init_master" />
-    <t_order_item_6 item_id="110600" order_id="1106" user_id="11" status="init_master" />
-    <t_order_item_6 item_id="110601" order_id="1106" user_id="11" status="init_master" />
-    <t_order_item_7 item_id="110700" order_id="1107" user_id="11" status="init_master" />
-    <t_order_item_7 item_id="110701" order_id="1107" user_id="11" status="init_master" />
-    <t_order_item_8 item_id="110800" order_id="1108" user_id="11" status="init_master" />
-    <t_order_item_8 item_id="110801" order_id="1108" user_id="11" status="init_master" />
-    <t_order_item_9 item_id="110900" order_id="1109" user_id="11" status="init_master" />
-    <t_order_item_9 item_id="110901" order_id="1109" user_id="11" status="init_master" />
+    <t_order_item_0 item_id="110000" order_id="1100" user_id="11" status="init" />
+    <t_order_item_0 item_id="110001" order_id="1100" user_id="11" status="init" />
+    <t_order_item_1 item_id="110100" order_id="1101" user_id="11" status="init" />
+    <t_order_item_1 item_id="110101" order_id="1101" user_id="11" status="init" />
+    <t_order_item_2 item_id="110200" order_id="1102" user_id="11" status="init" />
+    <t_order_item_2 item_id="110201" order_id="1102" user_id="11" status="init" />
+    <t_order_item_3 item_id="110300" order_id="1103" user_id="11" status="init" />
+    <t_order_item_3 item_id="110301" order_id="1103" user_id="11" status="init" />
+    <t_order_item_4 item_id="110400" order_id="1104" user_id="11" status="init" />
+    <t_order_item_4 item_id="110401" order_id="1104" user_id="11" status="init" />
+    <t_order_item_5 item_id="110500" order_id="1105" user_id="11" status="init" />
+    <t_order_item_5 item_id="110501" order_id="1105" user_id="11" status="init" />
+    <t_order_item_6 item_id="110600" order_id="1106" user_id="11" status="init" />
+    <t_order_item_6 item_id="110601" order_id="1106" user_id="11" status="init" />
+    <t_order_item_7 item_id="110700" order_id="1107" user_id="11" status="init" />
+    <t_order_item_7 item_id="110701" order_id="1107" user_id="11" status="init" />
+    <t_order_item_8 item_id="110800" order_id="1108" user_id="11" status="init" />
+    <t_order_item_8 item_id="110801" order_id="1108" user_id="11" status="init" />
+    <t_order_item_9 item_id="110900" order_id="1109" user_id="11" status="init" />
+    <t_order_item_9 item_id="110901" order_id="1109" user_id="11" status="init" />
     
-    <t_config id="1" status="init_master" />
+    <t_config id="1" status="init" />
 </dataset>
