+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/init/master_2.xml
@@ -1,35 +1,35 @@
 <dataset>
-    <t_order_0 order_id="1200" user_id="12" status="init_master" />
-    <t_order_1 order_id="1201" user_id="12" status="init_master" />
-    <t_order_2 order_id="1202" user_id="12" status="init_master" />
-    <t_order_3 order_id="1203" user_id="12" status="init_master" />
-    <t_order_4 order_id="1204" user_id="12" status="init_master" />
-    <t_order_5 order_id="1205" user_id="12" status="init_master" />
-    <t_order_6 order_id="1206" user_id="12" status="init_master" />
-    <t_order_7 order_id="1207" user_id="12" status="init_master" />
-    <t_order_8 order_id="1208" user_id="12" status="init_master" />
-    <t_order_9 order_id="1209" user_id="12" status="init_master" />
+    <t_order_0 order_id="1200" user_id="12" status="init" />
+    <t_order_1 order_id="1201" user_id="12" status="init" />
+    <t_order_2 order_id="1202" user_id="12" status="init" />
+    <t_order_3 order_id="1203" user_id="12" status="init" />
+    <t_order_4 order_id="1204" user_id="12" status="init" />
+    <t_order_5 order_id="1205" user_id="12" status="init" />
+    <t_order_6 order_id="1206" user_id="12" status="init" />
+    <t_order_7 order_id="1207" user_id="12" status="init" />
+    <t_order_8 order_id="1208" user_id="12" status="init" />
+    <t_order_9 order_id="1209" user_id="12" status="init" />
     
-    <t_order_item_0 item_id="120000" order_id="1200" user_id="12" status="init_master" />
-    <t_order_item_0 item_id="120001" order_id="1200" user_id="12" status="init_master" />
-    <t_order_item_1 item_id="120100" order_id="1201" user_id="12" status="init_master" />
-    <t_order_item_1 item_id="120101" order_id="1201" user_id="12" status="init_master" />
-    <t_order_item_2 item_id="120200" order_id="1202" user_id="12" status="init_master" />
-    <t_order_item_2 item_id="120201" order_id="1202" user_id="12" status="init_master" />
-    <t_order_item_3 item_id="120300" order_id="1203" user_id="12" status="init_master" />
-    <t_order_item_3 item_id="120301" order_id="1203" user_id="12" status="init_master" />
-    <t_order_item_4 item_id="120400" order_id="1204" user_id="12" status="init_master" />
-    <t_order_item_4 item_id="120401" order_id="1204" user_id="12" status="init_master" />
-    <t_order_item_5 item_id="120500" order_id="1205" user_id="12" status="init_master" />
-    <t_order_item_5 item_id="120501" order_id="1205" user_id="12" status="init_master" />
-    <t_order_item_6 item_id="120600" order_id="1206" user_id="12" status="init_master" />
-    <t_order_item_6 item_id="120601" order_id="1206" user_id="12" status="init_master" />
-    <t_order_item_7 item_id="120700" order_id="1207" user_id="12" status="init_master" />
-    <t_order_item_7 item_id="120701" order_id="1207" user_id="12" status="init_master" />
-    <t_order_item_8 item_id="120800" order_id="1208" user_id="12" status="init_master" />
-    <t_order_item_8 item_id="120801" order_id="1208" user_id="12" status="init_master" />
-    <t_order_item_9 item_id="120900" order_id="1209" user_id="12" status="init_master" />
-    <t_order_item_9 item_id="120901" order_id="1209" user_id="12" status="init_master" />
+    <t_order_item_0 item_id="120000" order_id="1200" user_id="12" status="init" />
+    <t_order_item_0 item_id="120001" order_id="1200" user_id="12" status="init" />
+    <t_order_item_1 item_id="120100" order_id="1201" user_id="12" status="init" />
+    <t_order_item_1 item_id="120101" order_id="1201" user_id="12" status="init" />
+    <t_order_item_2 item_id="120200" order_id="1202" user_id="12" status="init" />
+    <t_order_item_2 item_id="120201" order_id="1202" user_id="12" status="init" />
+    <t_order_item_3 item_id="120300" order_id="1203" user_id="12" status="init" />
+    <t_order_item_3 item_id="120301" order_id="1203" user_id="12" status="init" />
+    <t_order_item_4 item_id="120400" order_id="1204" user_id="12" status="init" />
+    <t_order_item_4 item_id="120401" order_id="1204" user_id="12" status="init" />
+    <t_order_item_5 item_id="120500" order_id="1205" user_id="12" status="init" />
+    <t_order_item_5 item_id="120501" order_id="1205" user_id="12" status="init" />
+    <t_order_item_6 item_id="120600" order_id="1206" user_id="12" status="init" />
+    <t_order_item_6 item_id="120601" order_id="1206" user_id="12" status="init" />
+    <t_order_item_7 item_id="120700" order_id="1207" user_id="12" status="init" />
+    <t_order_item_7 item_id="120701" order_id="1207" user_id="12" status="init" />
+    <t_order_item_8 item_id="120800" order_id="1208" user_id="12" status="init" />
+    <t_order_item_8 item_id="120801" order_id="1208" user_id="12" status="init" />
+    <t_order_item_9 item_id="120900" order_id="1209" user_id="12" status="init" />
+    <t_order_item_9 item_id="120901" order_id="1209" user_id="12" status="init" />
     
-    <t_config id="1" status="init_master" />
+    <t_config id="1" status="init" />
 </dataset>
