+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/init/master_3.xml
@@ -1,35 +1,35 @@
 <dataset>
-    <t_order_0 order_id="1300" user_id="13" status="init_master" />
-    <t_order_1 order_id="1301" user_id="13" status="init_master" />
-    <t_order_2 order_id="1302" user_id="13" status="init_master" />
-    <t_order_3 order_id="1303" user_id="13" status="init_master" />
-    <t_order_4 order_id="1304" user_id="13" status="init_master" />
-    <t_order_5 order_id="1305" user_id="13" status="init_master" />
-    <t_order_6 order_id="1306" user_id="13" status="init_master" />
-    <t_order_7 order_id="1307" user_id="13" status="init_master" />
-    <t_order_8 order_id="1308" user_id="13" status="init_master" />
-    <t_order_9 order_id="1309" user_id="13" status="init_master" />
+    <t_order_0 order_id="1300" user_id="13" status="init" />
+    <t_order_1 order_id="1301" user_id="13" status="init" />
+    <t_order_2 order_id="1302" user_id="13" status="init" />
+    <t_order_3 order_id="1303" user_id="13" status="init" />
+    <t_order_4 order_id="1304" user_id="13" status="init" />
+    <t_order_5 order_id="1305" user_id="13" status="init" />
+    <t_order_6 order_id="1306" user_id="13" status="init" />
+    <t_order_7 order_id="1307" user_id="13" status="init" />
+    <t_order_8 order_id="1308" user_id="13" status="init" />
+    <t_order_9 order_id="1309" user_id="13" status="init" />
     
-    <t_order_item_0 item_id="130000" order_id="1300" user_id="13" status="init_master" />
-    <t_order_item_0 item_id="130001" order_id="1300" user_id="13" status="init_master" />
-    <t_order_item_1 item_id="130100" order_id="1301" user_id="13" status="init_master" />
-    <t_order_item_1 item_id="130101" order_id="1301" user_id="13" status="init_master" />
-    <t_order_item_2 item_id="130200" order_id="1302" user_id="13" status="init_master" />
-    <t_order_item_2 item_id="130201" order_id="1302" user_id="13" status="init_master" />
-    <t_order_item_3 item_id="130300" order_id="1303" user_id="13" status="init_master" />
-    <t_order_item_3 item_id="130301" order_id="1303" user_id="13" status="init_master" />
-    <t_order_item_4 item_id="130400" order_id="1304" user_id="13" status="init_master" />
-    <t_order_item_4 item_id="130401" order_id="1304" user_id="13" status="init_master" />
-    <t_order_item_5 item_id="130500" order_id="1305" user_id="13" status="init_master" />
-    <t_order_item_5 item_id="130501" order_id="1305" user_id="13" status="init_master" />
-    <t_order_item_6 item_id="130600" order_id="1306" user_id="13" status="init_master" />
-    <t_order_item_6 item_id="130601" order_id="1306" user_id="13" status="init_master" />
-    <t_order_item_7 item_id="130700" order_id="1307" user_id="13" status="init_master" />
-    <t_order_item_7 item_id="130701" order_id="1307" user_id="13" status="init_master" />
-    <t_order_item_8 item_id="130800" order_id="1308" user_id="13" status="init_master" />
-    <t_order_item_8 item_id="130801" order_id="1308" user_id="13" status="init_master" />
-    <t_order_item_9 item_id="130900" order_id="1309" user_id="13" status="init_master" />
-    <t_order_item_9 item_id="130901" order_id="1309" user_id="13" status="init_master" />
+    <t_order_item_0 item_id="130000" order_id="1300" user_id="13" status="init" />
+    <t_order_item_0 item_id="130001" order_id="1300" user_id="13" status="init" />
+    <t_order_item_1 item_id="130100" order_id="1301" user_id="13" status="init" />
+    <t_order_item_1 item_id="130101" order_id="1301" user_id="13" status="init" />
+    <t_order_item_2 item_id="130200" order_id="1302" user_id="13" status="init" />
+    <t_order_item_2 item_id="130201" order_id="1302" user_id="13" status="init" />
+    <t_order_item_3 item_id="130300" order_id="1303" user_id="13" status="init" />
+    <t_order_item_3 item_id="130301" order_id="1303" user_id="13" status="init" />
+    <t_order_item_4 item_id="130400" order_id="1304" user_id="13" status="init" />
+    <t_order_item_4 item_id="130401" order_id="1304" user_id="13" status="init" />
+    <t_order_item_5 item_id="130500" order_id="1305" user_id="13" status="init" />
+    <t_order_item_5 item_id="130501" order_id="1305" user_id="13" status="init" />
+    <t_order_item_6 item_id="130600" order_id="1306" user_id="13" status="init" />
+    <t_order_item_6 item_id="130601" order_id="1306" user_id="13" status="init" />
+    <t_order_item_7 item_id="130700" order_id="1307" user_id="13" status="init" />
+    <t_order_item_7 item_id="130701" order_id="1307" user_id="13" status="init" />
+    <t_order_item_8 item_id="130800" order_id="1308" user_id="13" status="init" />
+    <t_order_item_8 item_id="130801" order_id="1308" user_id="13" status="init" />
+    <t_order_item_9 item_id="130900" order_id="1309" user_id="13" status="init" />
+    <t_order_item_9 item_id="130901" order_id="1309" user_id="13" status="init" />
     
-    <t_config id="1" status="init_master" />
+    <t_config id="1" status="init" />
 </dataset>
