+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/init/master_4.xml
@@ -1,35 +1,35 @@
 <dataset>
-    <t_order_0 order_id="1400" user_id="14" status="init_master" />
-    <t_order_1 order_id="1401" user_id="14" status="init_master" />
-    <t_order_2 order_id="1402" user_id="14" status="init_master" />
-    <t_order_3 order_id="1403" user_id="14" status="init_master" />
-    <t_order_4 order_id="1404" user_id="14" status="init_master" />
-    <t_order_5 order_id="1405" user_id="14" status="init_master" />
-    <t_order_6 order_id="1406" user_id="14" status="init_master" />
-    <t_order_7 order_id="1407" user_id="14" status="init_master" />
-    <t_order_8 order_id="1408" user_id="14" status="init_master" />
-    <t_order_9 order_id="1409" user_id="14" status="init_master" />
+    <t_order_0 order_id="1400" user_id="14" status="init" />
+    <t_order_1 order_id="1401" user_id="14" status="init" />
+    <t_order_2 order_id="1402" user_id="14" status="init" />
+    <t_order_3 order_id="1403" user_id="14" status="init" />
+    <t_order_4 order_id="1404" user_id="14" status="init" />
+    <t_order_5 order_id="1405" user_id="14" status="init" />
+    <t_order_6 order_id="1406" user_id="14" status="init" />
+    <t_order_7 order_id="1407" user_id="14" status="init" />
+    <t_order_8 order_id="1408" user_id="14" status="init" />
+    <t_order_9 order_id="1409" user_id="14" status="init" />
     
-    <t_order_item_0 item_id="140000" order_id="1400" user_id="14" status="init_master" />
-    <t_order_item_0 item_id="140001" order_id="1400" user_id="14" status="init_master" />
-    <t_order_item_1 item_id="140100" order_id="1401" user_id="14" status="init_master" />
-    <t_order_item_1 item_id="140101" order_id="1401" user_id="14" status="init_master" />
-    <t_order_item_2 item_id="140200" order_id="1402" user_id="14" status="init_master" />
-    <t_order_item_2 item_id="140201" order_id="1402" user_id="14" status="init_master" />
-    <t_order_item_3 item_id="140300" order_id="1403" user_id="14" status="init_master" />
-    <t_order_item_3 item_id="140301" order_id="1403" user_id="14" status="init_master" />
-    <t_order_item_4 item_id="140400" order_id="1404" user_id="14" status="init_master" />
-    <t_order_item_4 item_id="140401" order_id="1404" user_id="14" status="init_master" />
-    <t_order_item_5 item_id="140500" order_id="1405" user_id="14" status="init_master" />
-    <t_order_item_5 item_id="140501" order_id="1405" user_id="14" status="init_master" />
-    <t_order_item_6 item_id="140600" order_id="1406" user_id="14" status="init_master" />
-    <t_order_item_6 item_id="140601" order_id="1406" user_id="14" status="init_master" />
-    <t_order_item_7 item_id="140700" order_id="1407" user_id="14" status="init_master" />
-    <t_order_item_7 item_id="140701" order_id="1407" user_id="14" status="init_master" />
-    <t_order_item_8 item_id="140800" order_id="1408" user_id="14" status="init_master" />
-    <t_order_item_8 item_id="140801" order_id="1408" user_id="14" status="init_master" />
-    <t_order_item_9 item_id="140900" order_id="1409" user_id="14" status="init_master" />
-    <t_order_item_9 item_id="140901" order_id="1409" user_id="14" status="init_master" />
+    <t_order_item_0 item_id="140000" order_id="1400" user_id="14" status="init" />
+    <t_order_item_0 item_id="140001" order_id="1400" user_id="14" status="init" />
+    <t_order_item_1 item_id="140100" order_id="1401" user_id="14" status="init" />
+    <t_order_item_1 item_id="140101" order_id="1401" user_id="14" status="init" />
+    <t_order_item_2 item_id="140200" order_id="1402" user_id="14" status="init" />
+    <t_order_item_2 item_id="140201" order_id="1402" user_id="14" status="init" />
+    <t_order_item_3 item_id="140300" order_id="1403" user_id="14" status="init" />
+    <t_order_item_3 item_id="140301" order_id="1403" user_id="14" status="init" />
+    <t_order_item_4 item_id="140400" order_id="1404" user_id="14" status="init" />
+    <t_order_item_4 item_id="140401" order_id="1404" user_id="14" status="init" />
+    <t_order_item_5 item_id="140500" order_id="1405" user_id="14" status="init" />
+    <t_order_item_5 item_id="140501" order_id="1405" user_id="14" status="init" />
+    <t_order_item_6 item_id="140600" order_id="1406" user_id="14" status="init" />
+    <t_order_item_6 item_id="140601" order_id="1406" user_id="14" status="init" />
+    <t_order_item_7 item_id="140700" order_id="1407" user_id="14" status="init" />
+    <t_order_item_7 item_id="140701" order_id="1407" user_id="14" status="init" />
+    <t_order_item_8 item_id="140800" order_id="1408" user_id="14" status="init" />
+    <t_order_item_8 item_id="140801" order_id="1408" user_id="14" status="init" />
+    <t_order_item_9 item_id="140900" order_id="1409" user_id="14" status="init" />
+    <t_order_item_9 item_id="140901" order_id="1409" user_id="14" status="init" />
     
-    <t_config id="1" status="init_master" />
+    <t_config id="1" status="init" />
 </dataset>
