+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/init/master_5.xml
@@ -1,35 +1,35 @@
 <dataset>
-    <t_order_0 order_id="1500" user_id="15" status="init_master" />
-    <t_order_1 order_id="1501" user_id="15" status="init_master" />
-    <t_order_2 order_id="1502" user_id="15" status="init_master" />
-    <t_order_3 order_id="1503" user_id="15" status="init_master" />
-    <t_order_4 order_id="1504" user_id="15" status="init_master" />
-    <t_order_5 order_id="1505" user_id="15" status="init_master" />
-    <t_order_6 order_id="1506" user_id="15" status="init_master" />
-    <t_order_7 order_id="1507" user_id="15" status="init_master" />
-    <t_order_8 order_id="1508" user_id="15" status="init_master" />
-    <t_order_9 order_id="1509" user_id="15" status="init_master" />
+    <t_order_0 order_id="1500" user_id="15" status="init" />
+    <t_order_1 order_id="1501" user_id="15" status="init" />
+    <t_order_2 order_id="1502" user_id="15" status="init" />
+    <t_order_3 order_id="1503" user_id="15" status="init" />
+    <t_order_4 order_id="1504" user_id="15" status="init" />
+    <t_order_5 order_id="1505" user_id="15" status="init" />
+    <t_order_6 order_id="1506" user_id="15" status="init" />
+    <t_order_7 order_id="1507" user_id="15" status="init" />
+    <t_order_8 order_id="1508" user_id="15" status="init" />
+    <t_order_9 order_id="1509" user_id="15" status="init" />
     
-    <t_order_item_0 item_id="150000" order_id="1500" user_id="15" status="init_master" />
-    <t_order_item_0 item_id="150001" order_id="1500" user_id="15" status="init_master" />
-    <t_order_item_1 item_id="150100" order_id="1501" user_id="15" status="init_master" />
-    <t_order_item_1 item_id="150101" order_id="1501" user_id="15" status="init_master" />
-    <t_order_item_2 item_id="150200" order_id="1502" user_id="15" status="init_master" />
-    <t_order_item_2 item_id="150201" order_id="1502" user_id="15" status="init_master" />
-    <t_order_item_3 item_id="150300" order_id="1503" user_id="15" status="init_master" />
-    <t_order_item_3 item_id="150301" order_id="1503" user_id="15" status="init_master" />
-    <t_order_item_4 item_id="150400" order_id="1504" user_id="15" status="init_master" />
-    <t_order_item_4 item_id="150401" order_id="1504" user_id="15" status="init_master" />
-    <t_order_item_5 item_id="150500" order_id="1505" user_id="15" status="init_master" />
-    <t_order_item_5 item_id="150501" order_id="1505" user_id="15" status="init_master" />
-    <t_order_item_6 item_id="150600" order_id="1506" user_id="15" status="init_master" />
-    <t_order_item_6 item_id="150601" order_id="1506" user_id="15" status="init_master" />
-    <t_order_item_7 item_id="150700" order_id="1507" user_id="15" status="init_master" />
-    <t_order_item_7 item_id="150701" order_id="1507" user_id="15" status="init_master" />
-    <t_order_item_8 item_id="150800" order_id="1508" user_id="15" status="init_master" />
-    <t_order_item_8 item_id="150801" order_id="1508" user_id="15" status="init_master" />
-    <t_order_item_9 item_id="150900" order_id="1509" user_id="15" status="init_master" />
-    <t_order_item_9 item_id="150901" order_id="1509" user_id="15" status="init_master" />
+    <t_order_item_0 item_id="150000" order_id="1500" user_id="15" status="init" />
+    <t_order_item_0 item_id="150001" order_id="1500" user_id="15" status="init" />
+    <t_order_item_1 item_id="150100" order_id="1501" user_id="15" status="init" />
+    <t_order_item_1 item_id="150101" order_id="1501" user_id="15" status="init" />
+    <t_order_item_2 item_id="150200" order_id="1502" user_id="15" status="init" />
+    <t_order_item_2 item_id="150201" order_id="1502" user_id="15" status="init" />
+    <t_order_item_3 item_id="150300" order_id="1503" user_id="15" status="init" />
+    <t_order_item_3 item_id="150301" order_id="1503" user_id="15" status="init" />
+    <t_order_item_4 item_id="150400" order_id="1504" user_id="15" status="init" />
+    <t_order_item_4 item_id="150401" order_id="1504" user_id="15" status="init" />
+    <t_order_item_5 item_id="150500" order_id="1505" user_id="15" status="init" />
+    <t_order_item_5 item_id="150501" order_id="1505" user_id="15" status="init" />
+    <t_order_item_6 item_id="150600" order_id="1506" user_id="15" status="init" />
+    <t_order_item_6 item_id="150601" order_id="1506" user_id="15" status="init" />
+    <t_order_item_7 item_id="150700" order_id="1507" user_id="15" status="init" />
+    <t_order_item_7 item_id="150701" order_id="1507" user_id="15" status="init" />
+    <t_order_item_8 item_id="150800" order_id="1508" user_id="15" status="init" />
+    <t_order_item_8 item_id="150801" order_id="1508" user_id="15" status="init" />
+    <t_order_item_9 item_id="150900" order_id="1509" user_id="15" status="init" />
+    <t_order_item_9 item_id="150901" order_id="1509" user_id="15" status="init" />
     
-    <t_config id="1" status="init_master" />
+    <t_config id="1" status="init" />
 </dataset>
