+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/init/master_6.xml
@@ -1,35 +1,35 @@
 <dataset>
-    <t_order_0 order_id="1600" user_id="16" status="init_master" />
-    <t_order_1 order_id="1601" user_id="16" status="init_master" />
-    <t_order_2 order_id="1602" user_id="16" status="init_master" />
-    <t_order_3 order_id="1603" user_id="16" status="init_master" />
-    <t_order_4 order_id="1604" user_id="16" status="init_master" />
-    <t_order_5 order_id="1605" user_id="16" status="init_master" />
-    <t_order_6 order_id="1606" user_id="16" status="init_master" />
-    <t_order_7 order_id="1607" user_id="16" status="init_master" />
-    <t_order_8 order_id="1608" user_id="16" status="init_master" />
-    <t_order_9 order_id="1609" user_id="16" status="init_master" />
+    <t_order_0 order_id="1600" user_id="16" status="init" />
+    <t_order_1 order_id="1601" user_id="16" status="init" />
+    <t_order_2 order_id="1602" user_id="16" status="init" />
+    <t_order_3 order_id="1603" user_id="16" status="init" />
+    <t_order_4 order_id="1604" user_id="16" status="init" />
+    <t_order_5 order_id="1605" user_id="16" status="init" />
+    <t_order_6 order_id="1606" user_id="16" status="init" />
+    <t_order_7 order_id="1607" user_id="16" status="init" />
+    <t_order_8 order_id="1608" user_id="16" status="init" />
+    <t_order_9 order_id="1609" user_id="16" status="init" />
     
-    <t_order_item_0 item_id="160000" order_id="1600" user_id="16" status="init_master" />
-    <t_order_item_0 item_id="160001" order_id="1600" user_id="16" status="init_master" />
-    <t_order_item_1 item_id="160100" order_id="1601" user_id="16" status="init_master" />
-    <t_order_item_1 item_id="160101" order_id="1601" user_id="16" status="init_master" />
-    <t_order_item_2 item_id="160200" order_id="1602" user_id="16" status="init_master" />
-    <t_order_item_2 item_id="160201" order_id="1602" user_id="16" status="init_master" />
-    <t_order_item_3 item_id="160300" order_id="1603" user_id="16" status="init_master" />
-    <t_order_item_3 item_id="160301" order_id="1603" user_id="16" status="init_master" />
-    <t_order_item_4 item_id="160400" order_id="1604" user_id="16" status="init_master" />
-    <t_order_item_4 item_id="160401" order_id="1604" user_id="16" status="init_master" />
-    <t_order_item_5 item_id="160500" order_id="1605" user_id="16" status="init_master" />
-    <t_order_item_5 item_id="160501" order_id="1605" user_id="16" status="init_master" />
-    <t_order_item_6 item_id="160600" order_id="1606" user_id="16" status="init_master" />
-    <t_order_item_6 item_id="160601" order_id="1606" user_id="16" status="init_master" />
-    <t_order_item_7 item_id="160700" order_id="1607" user_id="16" status="init_master" />
-    <t_order_item_7 item_id="160701" order_id="1607" user_id="16" status="init_master" />
-    <t_order_item_8 item_id="160800" order_id="1608" user_id="16" status="init_master" />
-    <t_order_item_8 item_id="160801" order_id="1608" user_id="16" status="init_master" />
-    <t_order_item_9 item_id="160900" order_id="1609" user_id="16" status="init_master" />
-    <t_order_item_9 item_id="160901" order_id="1609" user_id="16" status="init_master" />
+    <t_order_item_0 item_id="160000" order_id="1600" user_id="16" status="init" />
+    <t_order_item_0 item_id="160001" order_id="1600" user_id="16" status="init" />
+    <t_order_item_1 item_id="160100" order_id="1601" user_id="16" status="init" />
+    <t_order_item_1 item_id="160101" order_id="1601" user_id="16" status="init" />
+    <t_order_item_2 item_id="160200" order_id="1602" user_id="16" status="init" />
+    <t_order_item_2 item_id="160201" order_id="1602" user_id="16" status="init" />
+    <t_order_item_3 item_id="160300" order_id="1603" user_id="16" status="init" />
+    <t_order_item_3 item_id="160301" order_id="1603" user_id="16" status="init" />
+    <t_order_item_4 item_id="160400" order_id="1604" user_id="16" status="init" />
+    <t_order_item_4 item_id="160401" order_id="1604" user_id="16" status="init" />
+    <t_order_item_5 item_id="160500" order_id="1605" user_id="16" status="init" />
+    <t_order_item_5 item_id="160501" order_id="1605" user_id="16" status="init" />
+    <t_order_item_6 item_id="160600" order_id="1606" user_id="16" status="init" />
+    <t_order_item_6 item_id="160601" order_id="1606" user_id="16" status="init" />
+    <t_order_item_7 item_id="160700" order_id="1607" user_id="16" status="init" />
+    <t_order_item_7 item_id="160701" order_id="1607" user_id="16" status="init" />
+    <t_order_item_8 item_id="160800" order_id="1608" user_id="16" status="init" />
+    <t_order_item_8 item_id="160801" order_id="1608" user_id="16" status="init" />
+    <t_order_item_9 item_id="160900" order_id="1609" user_id="16" status="init" />
+    <t_order_item_9 item_id="160901" order_id="1609" user_id="16" status="init" />
     
-    <t_config id="1" status="init_master" />
+    <t_config id="1" status="init" />
 </dataset>
