+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/init/master_7.xml
@@ -1,35 +1,35 @@
 <dataset>
-    <t_order_0 order_id="1700" user_id="17" status="init_master" />
-    <t_order_1 order_id="1701" user_id="17" status="init_master" />
-    <t_order_2 order_id="1702" user_id="17" status="init_master" />
-    <t_order_3 order_id="1703" user_id="17" status="init_master" />
-    <t_order_4 order_id="1704" user_id="17" status="init_master" />
-    <t_order_5 order_id="1705" user_id="17" status="init_master" />
-    <t_order_6 order_id="1706" user_id="17" status="init_master" />
-    <t_order_7 order_id="1707" user_id="17" status="init_master" />
-    <t_order_8 order_id="1708" user_id="17" status="init_master" />
-    <t_order_9 order_id="1709" user_id="17" status="init_master" />
+    <t_order_0 order_id="1700" user_id="17" status="init" />
+    <t_order_1 order_id="1701" user_id="17" status="init" />
+    <t_order_2 order_id="1702" user_id="17" status="init" />
+    <t_order_3 order_id="1703" user_id="17" status="init" />
+    <t_order_4 order_id="1704" user_id="17" status="init" />
+    <t_order_5 order_id="1705" user_id="17" status="init" />
+    <t_order_6 order_id="1706" user_id="17" status="init" />
+    <t_order_7 order_id="1707" user_id="17" status="init" />
+    <t_order_8 order_id="1708" user_id="17" status="init" />
+    <t_order_9 order_id="1709" user_id="17" status="init" />
     
-    <t_order_item_0 item_id="170000" order_id="1700" user_id="17" status="init_master" />
-    <t_order_item_0 item_id="170001" order_id="1700" user_id="17" status="init_master" />
-    <t_order_item_1 item_id="170100" order_id="1701" user_id="17" status="init_master" />
-    <t_order_item_1 item_id="170101" order_id="1701" user_id="17" status="init_master" />
-    <t_order_item_2 item_id="170200" order_id="1702" user_id="17" status="init_master" />
-    <t_order_item_2 item_id="170201" order_id="1702" user_id="17" status="init_master" />
-    <t_order_item_3 item_id="170300" order_id="1703" user_id="17" status="init_master" />
-    <t_order_item_3 item_id="170301" order_id="1703" user_id="17" status="init_master" />
-    <t_order_item_4 item_id="170400" order_id="1704" user_id="17" status="init_master" />
-    <t_order_item_4 item_id="170401" order_id="1704" user_id="17" status="init_master" />
-    <t_order_item_5 item_id="170500" order_id="1705" user_id="17" status="init_master" />
-    <t_order_item_5 item_id="170501" order_id="1705" user_id="17" status="init_master" />
-    <t_order_item_6 item_id="170600" order_id="1706" user_id="17" status="init_master" />
-    <t_order_item_6 item_id="170601" order_id="1706" user_id="17" status="init_master" />
-    <t_order_item_7 item_id="170700" order_id="1707" user_id="17" status="init_master" />
-    <t_order_item_7 item_id="170701" order_id="1707" user_id="17" status="init_master" />
-    <t_order_item_8 item_id="170800" order_id="1708" user_id="17" status="init_master" />
-    <t_order_item_8 item_id="170801" order_id="1708" user_id="17" status="init_master" />
-    <t_order_item_9 item_id="170900" order_id="1709" user_id="17" status="init_master" />
-    <t_order_item_9 item_id="170901" order_id="1709" user_id="17" status="init_master" />
+    <t_order_item_0 item_id="170000" order_id="1700" user_id="17" status="init" />
+    <t_order_item_0 item_id="170001" order_id="1700" user_id="17" status="init" />
+    <t_order_item_1 item_id="170100" order_id="1701" user_id="17" status="init" />
+    <t_order_item_1 item_id="170101" order_id="1701" user_id="17" status="init" />
+    <t_order_item_2 item_id="170200" order_id="1702" user_id="17" status="init" />
+    <t_order_item_2 item_id="170201" order_id="1702" user_id="17" status="init" />
+    <t_order_item_3 item_id="170300" order_id="1703" user_id="17" status="init" />
+    <t_order_item_3 item_id="170301" order_id="1703" user_id="17" status="init" />
+    <t_order_item_4 item_id="170400" order_id="1704" user_id="17" status="init" />
+    <t_order_item_4 item_id="170401" order_id="1704" user_id="17" status="init" />
+    <t_order_item_5 item_id="170500" order_id="1705" user_id="17" status="init" />
+    <t_order_item_5 item_id="170501" order_id="1705" user_id="17" status="init" />
+    <t_order_item_6 item_id="170600" order_id="1706" user_id="17" status="init" />
+    <t_order_item_6 item_id="170601" order_id="1706" user_id="17" status="init" />
+    <t_order_item_7 item_id="170700" order_id="1707" user_id="17" status="init" />
+    <t_order_item_7 item_id="170701" order_id="1707" user_id="17" status="init" />
+    <t_order_item_8 item_id="170800" order_id="1708" user_id="17" status="init" />
+    <t_order_item_8 item_id="170801" order_id="1708" user_id="17" status="init" />
+    <t_order_item_9 item_id="170900" order_id="1709" user_id="17" status="init" />
+    <t_order_item_9 item_id="170901" order_id="1709" user_id="17" status="init" />
     
-    <t_config id="1" status="init_master" />
+    <t_config id="1" status="init" />
 </dataset>
