+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/init/master_8.xml
@@ -1,35 +1,35 @@
 <dataset>
-    <t_order_0 order_id="1800" user_id="18" status="init_master" />
-    <t_order_1 order_id="1801" user_id="18" status="init_master" />
-    <t_order_2 order_id="1802" user_id="18" status="init_master" />
-    <t_order_3 order_id="1803" user_id="18" status="init_master" />
-    <t_order_4 order_id="1804" user_id="18" status="init_master" />
-    <t_order_5 order_id="1805" user_id="18" status="init_master" />
-    <t_order_6 order_id="1806" user_id="18" status="init_master" />
-    <t_order_7 order_id="1807" user_id="18" status="init_master" />
-    <t_order_8 order_id="1808" user_id="18" status="init_master" />
-    <t_order_9 order_id="1809" user_id="18" status="init_master" />
+    <t_order_0 order_id="1800" user_id="18" status="init" />
+    <t_order_1 order_id="1801" user_id="18" status="init" />
+    <t_order_2 order_id="1802" user_id="18" status="init" />
+    <t_order_3 order_id="1803" user_id="18" status="init" />
+    <t_order_4 order_id="1804" user_id="18" status="init" />
+    <t_order_5 order_id="1805" user_id="18" status="init" />
+    <t_order_6 order_id="1806" user_id="18" status="init" />
+    <t_order_7 order_id="1807" user_id="18" status="init" />
+    <t_order_8 order_id="1808" user_id="18" status="init" />
+    <t_order_9 order_id="1809" user_id="18" status="init" />
     
-    <t_order_item_0 item_id="180000" order_id="1800" user_id="18" status="init_master" />
-    <t_order_item_0 item_id="180001" order_id="1800" user_id="18" status="init_master" />
-    <t_order_item_1 item_id="180100" order_id="1801" user_id="18" status="init_master" />
-    <t_order_item_1 item_id="180101" order_id="1801" user_id="18" status="init_master" />
-    <t_order_item_2 item_id="180200" order_id="1802" user_id="18" status="init_master" />
-    <t_order_item_2 item_id="180201" order_id="1802" user_id="18" status="init_master" />
-    <t_order_item_3 item_id="180300" order_id="1803" user_id="18" status="init_master" />
-    <t_order_item_3 item_id="180301" order_id="1803" user_id="18" status="init_master" />
-    <t_order_item_4 item_id="180400" order_id="1804" user_id="18" status="init_master" />
-    <t_order_item_4 item_id="180401" order_id="1804" user_id="18" status="init_master" />
-    <t_order_item_5 item_id="180500" order_id="1805" user_id="18" status="init_master" />
-    <t_order_item_5 item_id="180501" order_id="1805" user_id="18" status="init_master" />
-    <t_order_item_6 item_id="180600" order_id="1806" user_id="18" status="init_master" />
-    <t_order_item_6 item_id="180601" order_id="1806" user_id="18" status="init_master" />
-    <t_order_item_7 item_id="180700" order_id="1807" user_id="18" status="init_master" />
-    <t_order_item_7 item_id="180701" order_id="1807" user_id="18" status="init_master" />
-    <t_order_item_8 item_id="180800" order_id="1808" user_id="18" status="init_master" />
-    <t_order_item_8 item_id="180801" order_id="1808" user_id="18" status="init_master" />
-    <t_order_item_9 item_id="180900" order_id="1809" user_id="18" status="init_master" />
-    <t_order_item_9 item_id="180901" order_id="1809" user_id="18" status="init_master" />
+    <t_order_item_0 item_id="180000" order_id="1800" user_id="18" status="init" />
+    <t_order_item_0 item_id="180001" order_id="1800" user_id="18" status="init" />
+    <t_order_item_1 item_id="180100" order_id="1801" user_id="18" status="init" />
+    <t_order_item_1 item_id="180101" order_id="1801" user_id="18" status="init" />
+    <t_order_item_2 item_id="180200" order_id="1802" user_id="18" status="init" />
+    <t_order_item_2 item_id="180201" order_id="1802" user_id="18" status="init" />
+    <t_order_item_3 item_id="180300" order_id="1803" user_id="18" status="init" />
+    <t_order_item_3 item_id="180301" order_id="1803" user_id="18" status="init" />
+    <t_order_item_4 item_id="180400" order_id="1804" user_id="18" status="init" />
+    <t_order_item_4 item_id="180401" order_id="1804" user_id="18" status="init" />
+    <t_order_item_5 item_id="180500" order_id="1805" user_id="18" status="init" />
+    <t_order_item_5 item_id="180501" order_id="1805" user_id="18" status="init" />
+    <t_order_item_6 item_id="180600" order_id="1806" user_id="18" status="init" />
+    <t_order_item_6 item_id="180601" order_id="1806" user_id="18" status="init" />
+    <t_order_item_7 item_id="180700" order_id="1807" user_id="18" status="init" />
+    <t_order_item_7 item_id="180701" order_id="1807" user_id="18" status="init" />
+    <t_order_item_8 item_id="180800" order_id="1808" user_id="18" status="init" />
+    <t_order_item_8 item_id="180801" order_id="1808" user_id="18" status="init" />
+    <t_order_item_9 item_id="180900" order_id="1809" user_id="18" status="init" />
+    <t_order_item_9 item_id="180901" order_id="1809" user_id="18" status="init" />
     
-    <t_config id="1" status="init_master" />
+    <t_config id="1" status="init" />
 </dataset>
