+++ sharding-jdbc-core/src/test/resources/integrate/dataset/sharding/masterslave/init/master_9.xml
@@ -1,35 +1,35 @@
 <dataset>
-    <t_order_0 order_id="1900" user_id="19" status="init_master" />
-    <t_order_1 order_id="1901" user_id="19" status="init_master" />
-    <t_order_2 order_id="1902" user_id="19" status="init_master" />
-    <t_order_3 order_id="1903" user_id="19" status="init_master" />
-    <t_order_4 order_id="1904" user_id="19" status="init_master" />
-    <t_order_5 order_id="1905" user_id="19" status="init_master" />
-    <t_order_6 order_id="1906" user_id="19" status="init_master" />
-    <t_order_7 order_id="1907" user_id="19" status="init_master" />
-    <t_order_8 order_id="1908" user_id="19" status="init_master" />
-    <t_order_9 order_id="1909" user_id="19" status="init_master" />
+    <t_order_0 order_id="1900" user_id="19" status="init" />
+    <t_order_1 order_id="1901" user_id="19" status="init" />
+    <t_order_2 order_id="1902" user_id="19" status="init" />
+    <t_order_3 order_id="1903" user_id="19" status="init" />
+    <t_order_4 order_id="1904" user_id="19" status="init" />
+    <t_order_5 order_id="1905" user_id="19" status="init" />
+    <t_order_6 order_id="1906" user_id="19" status="init" />
+    <t_order_7 order_id="1907" user_id="19" status="init" />
+    <t_order_8 order_id="1908" user_id="19" status="init" />
+    <t_order_9 order_id="1909" user_id="19" status="init" />
     
-    <t_order_item_0 item_id="190000" order_id="1900" user_id="19" status="init_master" />
-    <t_order_item_0 item_id="190001" order_id="1900" user_id="19" status="init_master" />
-    <t_order_item_1 item_id="190100" order_id="1901" user_id="19" status="init_master" />
-    <t_order_item_1 item_id="190101" order_id="1901" user_id="19" status="init_master" />
-    <t_order_item_2 item_id="190200" order_id="1902" user_id="19" status="init_master" />
-    <t_order_item_2 item_id="190201" order_id="1902" user_id="19" status="init_master" />
-    <t_order_item_3 item_id="190300" order_id="1903" user_id="19" status="init_master" />
-    <t_order_item_3 item_id="190301" order_id="1903" user_id="19" status="init_master" />
-    <t_order_item_4 item_id="190400" order_id="1904" user_id="19" status="init_master" />
-    <t_order_item_4 item_id="190401" order_id="1904" user_id="19" status="init_master" />
-    <t_order_item_5 item_id="190500" order_id="1905" user_id="19" status="init_master" />
-    <t_order_item_5 item_id="190501" order_id="1905" user_id="19" status="init_master" />
-    <t_order_item_6 item_id="190600" order_id="1906" user_id="19" status="init_master" />
-    <t_order_item_6 item_id="190601" order_id="1906" user_id="19" status="init_master" />
-    <t_order_item_7 item_id="190700" order_id="1907" user_id="19" status="init_master" />
-    <t_order_item_7 item_id="190701" order_id="1907" user_id="19" status="init_master" />
-    <t_order_item_8 item_id="190800" order_id="1908" user_id="19" status="init_master" />
-    <t_order_item_8 item_id="190801" order_id="1908" user_id="19" status="init_master" />
-    <t_order_item_9 item_id="190900" order_id="1909" user_id="19" status="init_master" />
-    <t_order_item_9 item_id="190901" order_id="1909" user_id="19" status="init_master" />
+    <t_order_item_0 item_id="190000" order_id="1900" user_id="19" status="init" />
+    <t_order_item_0 item_id="190001" order_id="1900" user_id="19" status="init" />
+    <t_order_item_1 item_id="190100" order_id="1901" user_id="19" status="init" />
+    <t_order_item_1 item_id="190101" order_id="1901" user_id="19" status="init" />
+    <t_order_item_2 item_id="190200" order_id="1902" user_id="19" status="init" />
+    <t_order_item_2 item_id="190201" order_id="1902" user_id="19" status="init" />
+    <t_order_item_3 item_id="190300" order_id="1903" user_id="19" status="init" />
+    <t_order_item_3 item_id="190301" order_id="1903" user_id="19" status="init" />
+    <t_order_item_4 item_id="190400" order_id="1904" user_id="19" status="init" />
+    <t_order_item_4 item_id="190401" order_id="1904" user_id="19" status="init" />
+    <t_order_item_5 item_id="190500" order_id="1905" user_id="19" status="init" />
+    <t_order_item_5 item_id="190501" order_id="1905" user_id="19" status="init" />
+    <t_order_item_6 item_id="190600" order_id="1906" user_id="19" status="init" />
+    <t_order_item_6 item_id="190601" order_id="1906" user_id="19" status="init" />
+    <t_order_item_7 item_id="190700" order_id="1907" user_id="19" status="init" />
+    <t_order_item_7 item_id="190701" order_id="1907" user_id="19" status="init" />
+    <t_order_item_8 item_id="190800" order_id="1908" user_id="19" status="init" />
+    <t_order_item_8 item_id="190801" order_id="1908" user_id="19" status="init" />
+    <t_order_item_9 item_id="190900" order_id="1909" user_id="19" status="init" />
+    <t_order_item_9 item_id="190901" order_id="1909" user_id="19" status="init" />
     
-    <t_config id="1" status="init_master" />
+    <t_config id="1" status="init" />
 </dataset>
